Fix -Wfloat-equal warnings about Ship::m_wheelTransition and m_wheelState.
This patch makes m_wheelTransition an int (it only ever takes the values
-1, 0, 1), and bumps the save format number.

+++ src/Ship.h
@@ -196,7 +196,7 @@ private:
 	bool m_testLanded;
 	float m_launchLockTimeout;
 	float m_wheelState;
-	float m_wheelTransition;
+	int m_wheelTransition;
 
 	vector3d m_thrusters;
 	vector3d m_angThrusters;
